requirements.txt
Django>=4.0,<5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
drf-yasg
python-decouple
django-cors-headers
gunicorn